+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Add support for the NTFS file system.
 - Add support for the ZFS file system.
 - Add support for the APFS file system.
+- Change input devices to be more responsive.
 
 ## 2.2.0